End of day
Since yesterday was technically just a planning day since we couldn't get the hardware, today could be called the first night, but it really is the second night. Anyways, the demo is tomorrow, so the code should be just about done. These switches were difficult to get the hang of, but in the end they were a ton of fun to work with.

The images are used as backgrounds for the switches, but now that I think of it, we could have used the line drawing tool to do the same thing without needing to flash anything to the switches... hmmm... all well it's just a prototype